Using Set Comparison Operator
This SQL script shows how the self join example problem can be solved using the set comparison clause 'some'. It uses the '=' comparator. As per the theory '= some' comparison in SQL is similar to used the 'in' operator which accounts for set membership. This solution does not use the cross product.